COM/Django/rest_api/requirements.txt
asgiref==3.10.0
Django==4.2
django-seed==0.3.1
djangorestframework==3.16.1
Faker==38.0.0
psycopg2==2.9.11
sqlparse==0.5.3
toposort==1.10
tzdata==2025.2
